fix: remove use state in each member analytics
