+++ wondrous-bot-admin/src/components/MembersAnalytics/index.tsx
@@ -1,14 +1,8 @@
 import { DataTitle, StyledUsername } from "./styles";
-import { useState } from "react";
-import Modal from "components/Shared/Modal";
-import { Box, Grid } from "@mui/material";
+import { Box } from "@mui/material";
 import { AddressComponent, IntegrationsComponent, NameComponent, XpComponent } from "./Common";
-import DiscordAnalytics from "./DiscordAnalytics";
-import Submissions from "./Submissions";
-import OnboardingSubmissions from "./OnboardingSubmissions";
 
-
-const GeneralInfo = ({ user }) => {
+export const GeneralInfo = ({ user }) => {
   const ITEMS = [
     {
       label: "Name",
@@ -52,56 +46,11 @@ const GeneralInfo = ({ user }) => {
   );
 };
 
-const MembersAnalytics = ({ value }) => {
-  const [activeCmtyUser, setActiveCmtyUser] = useState(null);
+const MembersAnalytics = ({ value, setValue }) => {
   const username = value?.username || value?.discordUsername || value?.telegramUsername || "N/A";
-
-  const handleUsernameClick = (cmtyUserId) => {
-    return setActiveCmtyUser(cmtyUserId);
-  };
-
-  const handleClose = () => {
-    return setActiveCmtyUser(null);
-  };
-
-  const CONFIG = [
-    {
-      id: "general-info",
-      component: () => <GeneralInfo user={value} />,
-    },
-    {
-      id: "messages-reactions",
-
-      component: () => <DiscordAnalytics user={value} />,
-    },
-    {
-      id: "onboarding",
-      component: () => <OnboardingSubmissions user={value} />,
-    },
-    {
-      id: "submissions",
-      component: () => <Submissions user={value} />,
-    },
-  ];
-
   return (
     <>
-      <Modal open={!!activeCmtyUser} onClose={handleClose} title={`${username}`} maxWidth={740}>
-        <Grid display="flex" gap="32px" flexDirection="column">
-          {activeCmtyUser
-            ? CONFIG.map((config, idx) => {
-                return <config.component key={config.id} />;
-              })
-            : null}
-        </Grid>
-      </Modal>
-      <StyledUsername
-        fontSize="14px"
-        lineHeight="14px"
-        textAlign="center"
-        width="100%"
-        onClick={() => handleUsernameClick(value?.id)}
-      >
+      <StyledUsername fontSize="14px" lineHeight="14px" textAlign="center" width="100%" onClick={() => setValue()}>
         {username}
       </StyledUsername>
     </>
